inference/pyproject.toml
[tool.poetry]
name = "duofinder"
version = "0.1.0"
description = ""
authors = ["seokulee <cd1004s@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
pytest = "^8.1.1"
pandas = "^2.2.1"
sqlmodel = "^0.0.16"
pydantic-settings = "^2.2.1"
nltk = "^3.8.1"
pillow = "^10.2.0"
uvicorn = "^0.28.0"
loguru = "^0.7.2"
torch = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
